Resetting the valgrind test to special
The valgrind test is presently broken having had a supp file written for Intel on the CX1 cluster, but now being run on Ubuntu LTS with gcc. Stephan plans to write a new supp file for this, but in the meantime as this is a known-broken (and will-fix) test, for non-Fluidity reasons, it's set to 'special' and will reset to 'long' once a new supp file is available.