Overhaul of longtests to fix paths
Many longtests assumed they would be run from within the fluidity sourcetree,
which is particularly unsafe given they reside in another repository. Resetting
tests here to look for fluidity on PATH as opposed to specifying a full path to
the binary; note this means testers need to make sure the version of fluidity
found first in PATH is the one they want to test with.